Carga ficheros del Volcado de la Base de Datos Nacional de Subvenciones (BDNS).

Carga un objeto la información correspondiente. Se puede cargar una versión previamente almacenada o bien descargar la última versión disponible.

carga_convocatorias() carga información de convocatorias.

carga_subvenciones() carga información de subvenciones.

La información se descarga en la carpeta especificada con el parámetro cache_dir. Además, almacena una copia ya procesada de cada archivo en formato .RData para reducir los tiempos de carga de datos en cada sesión independiente.

carga_convocatorias(
  cache_dir = rappdirs::user_data_dir("subvencionesES"),
  update = FALSE,
  rebuild_db = FALSE
)

carga_subvenciones(
  cache_dir = rappdirs::user_data_dir("subvencionesES"),
  update = FALSE,
  rebuild_db = FALSE
)

Argumentos

cache_dir

Directorio donde se quiere almacenar los datos. Por defecto rappdirs::user_data_dir("subvencionesES"). En este directorio se almacena tambíén la copia procesada en formato .RData

update

Variable lógica para descargar (TRUE) la versión más actual o cargar la copia local. Por defecto carga copia local (FALSE). Con la opción update = TRUE el paquete reescribe también el archivo .RData asociado.

rebuild_db

Vuelve a generar la base de datos local .RData a partir de los archivos csv ya descargados.

Valor

Un objeto en formato tibble con las cabeceras correctas.

Detalles

Las funciones devuelven las columnas según los nombres explicados en el issue #11.

Resultados de carga_convocatorias()

nombres_columnas	etiquetas_columnas

IDConv	IDConv

id	Código BDNS

mrr	MRR

convocanteN1	Administración

convocanteN2	Departamento

convocanteN3	Órgano

fechareg	Fecha de registro

titulo	Título de la convocatoria

bbreguladoras	URL de las BBRR

tituloleng	Título cooficial

verConcesiones	Ver concesiones

dummy1	dummy1

dummy2	dummy2

Resultados de carga_subvenciones()

nombres_columnas	etiquetas_columnas

ID	ID

IDConv	IDConvocatoria

convocanteN1	Administración

convocanteN2	Departamento

convocanteN3	Órgano

convocatoria	Convocatoria

bbreguladoras	URL de las BBRR

programa	Aplicación presupuestaria

fechaconc	Fecha de concesión

beneficiario	Beneficiario

importe	Importe

instrumento	Instrumento

ayudaequiv	Ayuda Equivalente

detalles	Detalles

proyecto	proyecto

sancion	sancion

numcov	numcov

Referencias

Github repo: https://github.com/JaimeObregon/subvenciones.

Autor-a

Diego Hernangómez, Javier Álvarez-Liébana.

Examples

if (FALSE) {
conv <- carga_convocatorias()
sub <- carga_subvenciones()
}