Extrae CIF del beneficiario

Extrae CIF del beneficiario

extrae_cif(x, col_benef = "beneficiario", append = TRUE)

Argumentos

x

Objeto inicial sobre el que se quiere ejecutar el split por CIF (Obtenido o derivado de carga_subvenciones()).

col_benef

Nombre de la columna donde se almacena el valor de beneficiario.

append

Variable lógica para añadir las nuevas columnas a x (TRUE) u obtener simplemente las columnas por separado (FALSE).

Valor

Un tibble con las columnas cif_beneficiario y nombre_beneficiario. Con la opción append = TRUE se devuelve el objeto x inicial con ambas columnas añadidas.

Detalles

Permite extraer el CIF y el nombre del beneficiario (sin el CIF) de un registro en concreto

Ver también

Examples

# Carga dataset ejemplo
sub <- carga_datasets_ejemplo("subvenciones")
#> ATENCION: Cargando datasets de ejemplo. No usar en produccion.
#> Cargando informacion almacenada en /home/runner/work/_temp/Library/subvencionesES/extdata/test/subvenciones.RData
#> Timestamp de la copia local: 2022-04-07 10:07:56 CEST
#> 
#> Para cargar la informacion completa usar 'carga_subvenciones()'

# Para carga completa usar:
# sub <- carga_subvenciones()

# Extrae CIF

library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

extrae_cif(sub) %>%
  select(IDConv, fechaconc, importe, cif_beneficiario, nombre_beneficiario)
#> # A tibble: 5,000 × 5
#>    IDConv fechaconc  importe cif_beneficiario nombre_beneficiario               
#>    <chr>  <date>       <dbl> <chr>            <chr>                             
#>  1 811910 2022-03-24    500  P4407400C        AYUNTAMIENTO DE CEDRILLAS         
#>  2 811910 2022-03-24    500  P4427500F        AYUNTAMIENTO DE VILLARQUEMADO     
#>  3 811910 2022-03-24    500  P4422100J        AYUNTAMIENTO DE SANTA EULALIA DEL…
#>  4 811910 2022-03-24    500  P4407600H        AYUNTAMIENTO DE CELLA             
#>  5 818661 2022-03-24  10000  G06330500        CLUB DEPORTIVO EXTREMADURA FEMENI…
#>  6 818652 2022-03-24   5000  P0601100A        ALMENDRALEJO, AYUNTAMIENTO        
#>  7 749877 2022-03-24   1500  G98545486        ASOCIACION CULTURAL ANDALUZA DE B…
#>  8 798146 2022-03-24   2000  G98943236        CLUB DE PATINATGE ARTISTIC RODAME…
#>  9 794733 2022-03-24 160000. G26324525        ASOC INVEST DESARR INNOV SECTOR A…
#> 10 794733 2022-03-24 160000. G26324525        ASOC INVEST DESARR INNOV SECTOR A…
#> # … with 4,990 more rows